Moves CohortData source and header files to /src and /include
