+++ src/ecodomain/Ground.h
@@ -25,7 +25,7 @@ using namespace std;
 #include "../lookup/CohortLookup.h"
 
 #include "../../include/BgcData.h"
-#include "../data/CohortData.h"
+#include "../../include/CohortData.h"
 #include "../data/FirData.h"
 #include "../data/RestartData.h"
