+++ src/snowsoil/Richards.h
@@ -8,7 +8,7 @@
 #include <cmath>
 #include <limits>
 
-#include "../data/CohortData.h"
+#include "../../include/CohortData.h"
 #include "../data/EnvData.h"
 
 #include "../ecodomain/layer/Layer.h"
