+++ src/snowsoil/Snow_Env.h
@@ -3,7 +3,7 @@
 
 #include "../inc/errorcode.h"
 
-#include "../data/CohortData.h"
+#include "../../include/CohortData.h"
 #include "../data/EnvData.h"
 #include "../data/RestartData.h"
