+++ src/snowsoil/Soil_Bgc.h
@@ -3,7 +3,7 @@
 
 #include "../inc/errorcode.h"
 
-#include "../data/CohortData.h"
+#include "../../include/CohortData.h"
 #include "../data/EnvData.h"
 #include "../data/FirData.h"
 #include "../../include/BgcData.h"
