+++ src/vegetation/Vegetation_Bgc.h
@@ -3,7 +3,7 @@
 #include "../lookup/CohortLookup.h"
 #include "../../include/ModelData.h"
 
-#include "../data/CohortData.h"
+#include "../../include/CohortData.h"
 #include "../data/EnvData.h"
 #include "../data/FirData.h"
 #include "../../include/BgcData.h"
